修复跨版本渲染错误

+++ src/main/java/net/minecraft/client/renderer/chunk/SetVisibility.java
@@ -80,7 +80,7 @@ public class SetVisibility
 
     private boolean getBit(int p_getBit_1_)
     {
-        return (this.bits & (long)(1 << p_getBit_1_)) != 0L;
+        return (this.bits & 1L << p_getBit_1_) != 0L;
     }
 
     private void setBit(int p_setBit_1_, boolean p_setBit_2_)
@@ -97,11 +97,11 @@ public class SetVisibility
 
     private void setBit(int p_setBit_1_)
     {
-        this.bits |= 1 << p_setBit_1_;
+        this.bits |= 1L << p_setBit_1_;
     }
 
     private void clearBit(int p_clearBit_1_)
     {
-        this.bits &= ~(1 << p_clearBit_1_);
+        this.bits &= ~(1L << p_clearBit_1_);
     }
 }
